[data.search] Move SearchSource to common directory. (#77823)
