+++ src/plugins/data/common/search/index.ts
@@ -20,6 +20,6 @@
 export * from './aggs';
 export * from './es_search';
 export * from './expressions';
+export * from './search_source';
 export * from './tabify';
 export * from './types';
-export * from './es_search';
