+++ src/plugins/data/common/search/search_source/create_search_source.ts
