+++ src/plugins/data/common/search/search_source/fetch/get_search_params.test.ts
@@ -17,20 +17,21 @@
  * under the License.
  */
 
-import { SearchResponse } from 'elasticsearch';
-import { KbnError } from '../../../../kibana_utils/common';
-import { SearchError } from './types';
+import { UI_SETTINGS } from '../../../constants';
+import { GetConfigFn } from '../../../types';
+import { getSearchParams } from './get_search_params';
 
-/**
- * Request Failure - When an entire multi request fails
- * @param {Error} err - the Error that came back
- * @param {Object} resp - optional HTTP response
- */
-export class RequestFailure extends KbnError {
-  public resp?: SearchResponse<any>;
-  constructor(err: SearchError | null = null, resp?: SearchResponse<any>) {
-    super(`Request to Elasticsearch failed: ${JSON.stringify(resp || err?.message)}`);
-
-    this.resp = resp;
-  }
+function getConfigStub(config: any = {}): GetConfigFn {
+  return (key) => config[key];
 }
+
+describe('getSearchParams', () => {
+  test('includes custom preference', () => {
+    const config = getConfigStub({
+      [UI_SETTINGS.COURIER_SET_REQUEST_PREFERENCE]: 'custom',
+      [UI_SETTINGS.COURIER_CUSTOM_REQUEST_PREFERENCE]: 'aaa',
+    });
+    const searchParams = getSearchParams(config);
+    expect(searchParams.preference).toBe('aaa');
+  });
+});
