+++ src/plugins/data/common/search/search_source/fetch/get_search_params.ts
@@ -17,7 +17,9 @@
  * under the License.
  */
 
-import { UI_SETTINGS, ISearchRequestParams, GetConfigFn } from '../../../common';
+import { UI_SETTINGS } from '../../../constants';
+import { GetConfigFn } from '../../../types';
+import { ISearchRequestParams } from '../../index';
 import { SearchRequest } from './types';
 
 const sessionId = Date.now();
