+++ src/plugins/data/common/search/search_source/fetch/request_error.ts
@@ -17,4 +17,20 @@
  * under the License.
  */
 
-export * from './call_msearch';
+import { SearchResponse } from 'elasticsearch';
+import { KbnError } from '../../../../../kibana_utils/common';
+import { SearchError } from './types';
+
+/**
+ * Request Failure - When an entire multi request fails
+ * @param {Error} err - the Error that came back
+ * @param {Object} resp - optional HTTP response
+ */
+export class RequestFailure extends KbnError {
+  public resp?: SearchResponse<any>;
+  constructor(err: SearchError | null = null, resp?: SearchResponse<any>) {
+    super(`Request to Elasticsearch failed: ${JSON.stringify(resp || err?.message)}`);
+
+    this.resp = resp;
+  }
+}
