+++ src/plugins/data/common/search/search_source/legacy/default_search_strategy.ts
