+++ src/plugins/data/common/search/search_source/legacy/fetch_soon.ts
@@ -18,10 +18,10 @@
  */
 
 import { SearchResponse } from 'elasticsearch';
+import { UI_SETTINGS } from '../../../constants';
+import { FetchHandlers, SearchRequest } from '../fetch';
+import { ISearchOptions } from '../../index';
 import { callClient } from './call_client';
-import { FetchHandlers } from '../fetch/types';
-import { SearchRequest } from '../index';
-import { UI_SETTINGS, ISearchOptions } from '../../../common';
 
 /**
  * This function introduces a slight delay in the request process to allow multiple requests to queue
