+++ src/plugins/data/common/search/search_source/legacy/index.ts
@@ -17,31 +17,5 @@
  * under the License.
  */
 
-import { BehaviorSubject } from 'rxjs';
-import { SearchResponse } from 'elasticsearch';
-import { FetchHandlers } from '../fetch';
-import { SearchRequest } from '..';
-
-// @internal
-export interface LegacyFetchHandlers {
-  callMsearch: (params: {
-    body: SearchRequest;
-    signal: AbortSignal;
-  }) => Promise<Array<SearchResponse<any>>>;
-  loadingCount$: BehaviorSubject<number>;
-}
-
-export interface SearchStrategySearchParams extends FetchHandlers {
-  searchRequests: SearchRequest[];
-}
-
-// @deprecated
-export interface SearchStrategyProvider {
-  id: string;
-  search: (params: SearchStrategySearchParams) => SearchStrategyResponse;
-}
-
-export interface SearchStrategyResponse<T = any> {
-  searching: Promise<Array<SearchResponse<T>>>;
-  abort: () => void;
-}
+export { fetchSoon } from './fetch_soon';
+export * from './types';
