+++ src/plugins/data/common/search/search_source/migrate_legacy_query.ts
@@ -17,4 +17,21 @@
  * under the License.
  */
 
-export * from './call_msearch';
+import { has } from 'lodash';
+import { Query } from '../../query/types';
+
+/**
+ * Creates a standardized query object from old queries that were either strings or pure ES query DSL
+ *
+ * @param query - a legacy query, what used to be stored in SearchSource's query property
+ * @return Object
+ */
+
+export function migrateLegacyQuery(query: Query | { [key: string]: any } | string): Query {
+  // Lucene was the only option before, so language-less queries are all lucene
+  if (!has(query, 'language')) {
+    return { query, language: 'lucene' };
+  }
+
+  return query as Query;
+}
