+++ src/plugins/data/public/search/fetch/index.ts
@@ -17,8 +17,4 @@
  * under the License.
  */
 
-export * from './types';
-export { getSearchParams, getSearchParamsFromRequest, getPreference } from './get_search_params';
-
-export { RequestFailure } from './request_error';
 export { handleResponse } from './handle_response';
