+++ src/plugins/data/public/search/legacy/call_msearch.ts
@@ -18,7 +18,7 @@
  */
 
 import { HttpStart } from 'src/core/public';
-import { LegacyFetchHandlers } from './types';
+import { LegacyFetchHandlers } from '../../../common/search/search_source';
 
 /**
  * Wrapper for calling the internal msearch endpoint from the client.
