+++ src/plugins/data/public/search/mocks.ts
@@ -19,9 +19,9 @@
 
 import { searchAggsSetupMock, searchAggsStartMock } from './aggs/mocks';
 import { ISearchSetup, ISearchStart } from './types';
-import { searchSourceMock, createSearchSourceMock } from './search_source/mocks';
+import { searchSourceMock, createSearchSourceMock } from '../../common/search/search_source/mocks';
 
-export * from './search_source/mocks';
+export * from '../../common/search/search_source/mocks';
 
 function createSetupContract(): jest.Mocked<ISearchSetup> {
   return {
