+++ src/plugins/data/public/search/types.ts
@@ -17,38 +17,18 @@
  * under the License.
  */
 
-import { Observable } from 'rxjs';
 import { PackageInfo } from 'kibana/server';
 import { ISearchInterceptor } from './search_interceptor';
-import { ISearchSource, SearchSourceFields } from './search_source';
 import { SearchUsageCollector } from './collectors';
 import { AggsSetup, AggsSetupDependencies, AggsStartDependencies, AggsStart } from './aggs';
-import {
-  IKibanaSearchRequest,
-  IKibanaSearchResponse,
-  IEsSearchRequest,
-  IEsSearchResponse,
-  ISearchOptions,
-} from '../../common/search';
+import { ISearchGeneric, ISearchSource, SearchSourceFields } from '../../common/search';
 import { IndexPatternsContract } from '../../common/index_patterns/index_patterns';
 import { UsageCollectionSetup } from '../../../usage_collection/public';
-
-export type ISearch = (
-  request: IKibanaSearchRequest,
-  options?: ISearchOptions
-) => Observable<IKibanaSearchResponse>;
-
-export type ISearchGeneric = <
-  SearchStrategyRequest extends IEsSearchRequest = IEsSearchRequest,
-  SearchStrategyResponse extends IEsSearchResponse = IEsSearchResponse
->(
-  request: SearchStrategyRequest,
-  options?: ISearchOptions
-) => Observable<SearchStrategyResponse>;
 
 export interface SearchEnhancements {
   searchInterceptor: ISearchInterceptor;
 }
+
 /**
  * The setup contract exposed by the Search plugin exposes the search strategy extension
  * point.
